requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
crispy-bootstrap5==2025.6
dj-config-url==0.1.1
django==6.0.1
django-allauth==65.14.0
django-cors-headers==4.9.0
django-crispy-forms==2.5
django-environ==0.12.0
django-htmx==1.27.0
django-modeltranslation==0.19.19
django-tinymce==5.0.0
djangorestframework==3.16.1
gunicorn==24.1.1
idna==3.11
packaging==26.0
pillow==12.1.0
psycopg2-binary==2.9.11
python-decouple==3.8
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.5
stripe==14.3.0
typing-extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.11.0
